requirements.txt
appdirs==1.4.2
dj-database-url==0.4.2
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
django-leaflet==0.21.0
djangorestframework==3.5.4
djangorestframework-gis==0.11
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
whitenoise==3.3.0
